refactor(data-model): load fonts on demand instead of open-time preload
Remove the FONT conversion stage, AcDbFontLoader options, and
AcDbFontNameCollector so open no longer preloads fonts. Also drop the
example DXF converter benchmark UI.
