+++ packages/data-model/__tests__/AcDbDatabase.spec.ts
@@ -84,138 +84,6 @@ describe('AcDbDatabase', () => {
     expect(db.generateHandle()).toBe('FFFF')
   })
 
-  it('loads multiple fonts in one batch during FONT END', async () => {
-    const db = new AcDbDatabase()
-    const fileType = 'test-font-load-batch'
-    const fontLoader = {
-      load: jest.fn().mockResolvedValue(undefined),
-      getAvaiableFonts: jest.fn().mockResolvedValue([])
-    }
-    const converter = {
-      read: jest.fn(
-        async (
-          _data: ArrayBuffer,
-          _db: AcDbDatabase,
-          _minimumChunkSize: number,
-          progress?: (
-            percentage: number,
-            stage: string,
-            stageStatus: string,
-            data?: unknown
-          ) => Promise<void>
-        ) => {
-          if (progress) {
-            await progress(5, 'FONT', 'END', ['arial', 'simsun', 'hztxt'])
-          }
-        }
-      )
-    }
-    const manager = AcDbDatabaseConverterManager.instance
-    manager.register(fileType, converter as never)
-
-    try {
-      await db.read(
-        new ArrayBuffer(0),
-        { fontLoader, readOnly: true },
-        fileType
-      )
-      expect(fontLoader.load).toHaveBeenCalledTimes(1)
-      expect(fontLoader.load).toHaveBeenCalledWith([
-        'arial',
-        'simsun',
-        'hztxt'
-      ])
-    } finally {
-      manager.unregister(fileType)
-    }
-  })
-
-  it('continues reading when font loading fails by default', async () => {
-    const db = new AcDbDatabase()
-    const fileType = 'test-font-load'
-    const fontLoader = {
-      load: jest.fn().mockRejectedValue(new Error('Failed to fetch')),
-      getAvaiableFonts: jest.fn().mockResolvedValue([])
-    }
-    const converter = {
-      read: jest.fn(
-        async (
-          _data: ArrayBuffer,
-          _db: AcDbDatabase,
-          _minimumChunkSize: number,
-          progress?: (
-            percentage: number,
-            stage: string,
-            stageStatus: string,
-            data?: unknown
-          ) => Promise<void>
-        ) => {
-          if (progress) {
-            await progress(5, 'FONT', 'END', ['arial'])
-          }
-        }
-      )
-    }
-    const manager = AcDbDatabaseConverterManager.instance
-    manager.register(fileType, converter as never)
-
-    const warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {})
-
-    try {
-      await expect(
-        db.read(new ArrayBuffer(0), { fontLoader, readOnly: true }, fileType)
-      ).resolves.toBeUndefined()
-      expect(fontLoader.load).toHaveBeenCalledWith(['arial'])
-      expect(warnSpy).toHaveBeenCalled()
-    } finally {
-      warnSpy.mockRestore()
-      manager.unregister(fileType)
-    }
-  })
-
-  it('aborts reading when font loading fails and failOnFontLoadError is true', async () => {
-    const db = new AcDbDatabase()
-    const fileType = 'test-font-load-strict'
-    const loadError = new Error('Failed to fetch')
-    const fontLoader = {
-      load: jest.fn().mockRejectedValue(loadError),
-      getAvaiableFonts: jest.fn().mockResolvedValue([])
-    }
-    const converter = {
-      read: jest.fn(
-        async (
-          _data: ArrayBuffer,
-          _db: AcDbDatabase,
-          _minimumChunkSize: number,
-          progress?: (
-            percentage: number,
-            stage: string,
-            stageStatus: string,
-            data?: unknown
-          ) => Promise<void>
-        ) => {
-          if (progress) {
-            await progress(5, 'FONT', 'END', ['arial'])
-          }
-        }
-      )
-    }
-    const manager = AcDbDatabaseConverterManager.instance
-    manager.register(fileType, converter as never)
-
-    try {
-      await expect(
-        db.read(
-          new ArrayBuffer(0),
-          { fontLoader, readOnly: true, failOnFontLoadError: true },
-          fileType
-        )
-      ).rejects.toThrow('Failed to fetch')
-    } finally {
-      manager.unregister(fileType)
-    }
-  })
-
   it('exposes worker OOM failures via lastOpenError, openFailed, and openProgress', async () => {
     const db = new AcDbDatabase()
     const fileType = 'test-open-failure'
