+++ packages/data-model/src/converter/index.ts
@@ -1,12 +1,4 @@
 export { AcDbBatchProcessing } from './AcDbBatchProcessing'
-export { AcDbFontNameCollector } from './AcDbFontNameCollector'
-export type {
-  AcDbFontNameCollectorAdapter,
-  AcDbFontNameCollectorEntities,
-  AcDbFontNameCollectorEntityFontInfo,
-  AcDbFontNameCollectorOptions,
-  AcDbFontNameCollectorStyleEntry
-} from './AcDbFontNameCollector'
 export { AcDbRegenerator } from './AcDbRegenerator'
 export {
   AcDbBaseWorker,
