+++ packages/example/package.json
@@ -41,7 +41,6 @@
   },
   "dependencies": {
     "@mlightcad/data-model": "workspace:*",
-    "@mlightcad/dxf-json-converter": "^1.12.0",
     "@mlightcad/libredwg-converter": "workspace:*"
   },
   "devDependencies": {
