+++ packages/libredwg-converter/__tests__/AcDbLibreDwgConverter.spec.ts
@@ -3,241 +3,12 @@ import { AcDbDatabase } from '@mlightcad/data-model'
 import { AcDbLibreDwgConverter } from '../src/AcDbLibreDwgConverter'
 
 class TestLibreDwgConverter extends AcDbLibreDwgConverter {
-  getFontsPublic(dwg: any) {
-    return this.getFonts(dwg)
-  }
-
   processHeaderPublic(model: any, db: AcDbDatabase) {
     return this.processHeader(model, db)
   }
 }
 
 describe('AcDbLibreDwgConverter', () => {
-  it('collects fonts from style table, mtext and nested inserts', () => {
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'A' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: 'A',
-              font: 'Arial.ttf',
-              bigFont: 'Bigfont.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        BLOCK_RECORD: {
-          entries: [
-            {
-              name: 'B1',
-              entities: [{ type: 'TEXT', styleName: 'A' }]
-            }
-          ]
-        }
-      },
-      entities: [
-        { type: 'TEXT', styleName: 'A' },
-        { type: 'MTEXT', styleName: 'A', text: '{\\fCustom|b0|i0;Hello}' },
-        { type: 'INSERT', name: 'B1' }
-      ]
-    })
-
-    expect(fonts).toEqual(
-      expect.arrayContaining(['arial', 'bigfont', 'custom'])
-    )
-  })
-
-  it('collects fonts for entities with missing style via TEXTSTYLE fallback', () => {
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'Standard' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: 'Standard',
-              font: 'txt.shx',
-              standardFlag: 0
-            },
-            {
-              name: 'Named',
-              font: 'simplex.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        BLOCK_RECORD: { entries: [] }
-      },
-      entities: [
-        { type: 'MTEXT', text: 'Hello' },
-        { type: 'TEXT', styleName: 'Named' },
-        { type: 'ATTRIB', styleName: 'Named' }
-      ]
-    })
-
-    expect(fonts).toEqual(expect.arrayContaining(['txt', 'simplex']))
-  })
-
-  it('collects shape-definition fonts and resolves MLeader text styles', () => {
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'Standard' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: '',
-              font: 'tecosymbol.shx',
-              standardFlag: 1
-            },
-            {
-              name: 'LeaderStyle',
-              font: 'romans.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        BLOCK_RECORD: { entries: [] }
-      },
-      entities: [
-        {
-          type: 'MULTILEADER',
-          textContent: '{\\fInline|b0;Note}',
-          textStyleName: 'LeaderStyle'
-        }
-      ]
-    })
-
-    expect(fonts).toEqual(
-      expect.arrayContaining(['tecosymbol', 'romans', 'inline'])
-    )
-  })
-
-  it('collects fonts from block definitions not referenced by model space inserts', () => {
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'Standard' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: 'Standard',
-              font: 'txt.shx',
-              bigFont: 'gbcbig.shx',
-              standardFlag: 0
-            },
-            {
-              name: 'Romans',
-              font: 'romans.shx',
-              standardFlag: 0
-            },
-            {
-              name: 'Unused',
-              font: 'italic.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        BLOCK_RECORD: {
-          entries: [
-            {
-              name: 'TITLE',
-              entities: [{ type: 'TEXT', styleName: 'Romans' }]
-            }
-          ]
-        }
-      },
-      entities: []
-    })
-
-    // Unused named STYLE fonts are skipped; only used block styles are kept.
-    expect(fonts).toEqual(expect.arrayContaining(['romans']))
-    expect(fonts).not.toContain('txt')
-    expect(fonts).not.toContain('gbcbig')
-    expect(fonts).not.toContain('italic')
-  })
-
-  it('does not throw when a BLOCK_RECORD has more entities than apply arg limits', () => {
-    // Engines typically fail around ~65k–125k args for Function.prototype.apply /
-    // array spread used as push(...entities). Use well above that.
-    const LARGE = 150_000
-    const entities = new Array(LARGE)
-    for (let i = 0; i < LARGE; i++) {
-      entities[i] = { type: 'LINE' }
-    }
-    entities[0] = { type: 'TEXT', styleName: 'Romans' }
-    entities[LARGE - 1] = {
-      type: 'MTEXT',
-      styleName: 'Romans',
-      text: '{\\fCustom|b0;Hi}'
-    }
-
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'Standard' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: 'Standard',
-              font: 'txt.shx',
-              standardFlag: 0
-            },
-            {
-              name: 'Romans',
-              font: 'romans.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        BLOCK_RECORD: {
-          entries: [{ name: 'DENSE', entities }]
-        }
-      },
-      entities: []
-    })
-
-    expect(fonts).toEqual(expect.arrayContaining(['romans', 'custom']))
-  })
-
-  it('collects fonts from tolerance entities via dim style text style and inline fonts', () => {
-    const converter = new TestLibreDwgConverter({ useWorker: false })
-    const fonts = converter.getFontsPublic({
-      header: { TEXTSTYLE: 'Standard' },
-      tables: {
-        STYLE: {
-          entries: [
-            {
-              name: 'Standard',
-              font: 'txt.shx',
-              standardFlag: 0
-            },
-            {
-              name: 'DimText',
-              font: 'romans.shx',
-              standardFlag: 0
-            }
-          ]
-        },
-        DIMSTYLE: {
-          entries: [{ name: 'Standard', DIMTXSTY: 'DimText' }]
-        },
-        BLOCK_RECORD: { entries: [] }
-      },
-      entities: [
-        {
-          type: 'TOLERANCE',
-          styleName: 'Standard',
-          text: '{\\Fgdt.shx|b0|i0;c134|p6;j}|0.05|A|'
-        }
-      ]
-    })
-
-    expect(fonts).toEqual(expect.arrayContaining(['romans', 'gdt.shx']))
-  })
-
   it('sets thumbnailImage from model thumbnailImage bytes', () => {
     const db = new AcDbDatabase()
     const converter = new TestLibreDwgConverter({ useWorker: false })
